refactor: move Input into SharedComponents
