+++ src/components/DownloadAndShareComponents/ResourcePageLayout/ResourcePageLayout.tsx
@@ -23,7 +23,7 @@ import FieldError from "@/components/FormFields/FieldError";
 import Box from "@/components/SharedComponents/Box";
 import Checkbox from "@/components/SharedComponents/Checkbox";
 import Flex from "@/components/SharedComponents/Flex";
-import Input from "@/components/Input";
+import Input from "@/components/SharedComponents/Input";
 import OakLink from "@/components/OakLink";
 import Icon from "@/components/SharedComponents/Icon";
