+++ src/components/SharedComponents/Input/Input.stories.tsx
