+++ src/components/SharedComponents/Input/Input.test.tsx
@@ -1,8 +1,9 @@
 import { screen } from "@testing-library/react";
 
-import renderWithTheme from "../../__tests__/__helpers__/renderWithTheme";
-
 import Input from "./Input";
+
+import renderWithTheme from "@/__tests__/__helpers__/renderWithTheme";
+
 
 describe("Input", () => {
   it("renders an input", () => {
